Strings, Files, and Improved AnnotationProcessor Annotation! Oh My!
- Added StringUtil for building strings from collections with a
separator between them

- Added FileUtil for reading/writing files once you already have a
Reader/Writer setup

- Added createFile, getExistingFile, getFileReader, and getFileWriter to
AnnotationUtil, as they function differently than a typical file (using
FileObject instead of a normal File too)

- Changed AnnotationProcessorProcessor to read the existing file and add
to it, rather than wiping it clean every time. The reason for this is
that when adding a new @AnnotationProcessor annotation causes it to
remake the file with just the new processor, and not include the old
list. - This does mean that the project needs cleaned in order to
generate the file if any processors are removed