Improved File Format and Made Stuff Final in Util
- Util classes (AutoCloseableUtil, BooleanUtil, ExceptionUtil, FileUtil,
ListUtil, and StringUtil) were made final, as they're not meant to be
extended or anything - also their methods are final if they weren't
before
- File Format now has a save function in it, to be called after the
particular file format converts its input type into the correct Nodes.
It checks that the Nodes are of the right format, then saves the file,
then checks the file's format is correct. Check 1 checks that the file
format obeyed its format, and Check 2 checks that the save and load
functionality is working properly