Added Event package from KBWCO
- Events were implemented first in KBWCO (1000 Blank White Cards Online,
a project I'm working on that will utilize these utils)
- Event is my own event class that just holds what type of event it is
as a string and the message for the event
- EventListener is used to listen for events and properly handle them in
whatever classes implement the listener. It is generic to be extended in
more specific EventListeners to handle more specific events (E.g.
ViewChangeEventListener to listen for ViewChangeEvents - which will be
in a commit soon)
- EventHandler is used to create events and send them out to the
appropriate listeners. It also is meant to be extended by other classes
for more specific events.