TadukooView from KBWCO
- KBWCO is in the process of switching to using a more generic View
framework. This is the progress so far (though slightly ahead of KBWCO
due to removing some KBWCO-specific stuff)

==View and ViewBase==
- ViewBase is to be extended in any specific program to be used as the
base framework class used to initialize the start of the program and
handle basically everything
- ViewBase ultimately uses Views (that are stored in the Context object)
that control a particular View (menu, game screen, etc.) in order to
control specifics for their particular menu/game screen/etc.
- ViewBase implements ViewChangeEventListener so it can properly close
or init Views as needed

==ViewChangeEvents==
- ViewChangeEvent is an extension of Event that is specific to the View
stack changing (opening/closing/changing Views)
- ViewChangeEventHandler and ViewChangeEventListener ought to be
extended in an actual program in order to be able to handle more
specific events, such as an OpenLoginEvent or similar

==Drawable, Clickable, and TextInput! Oh My!==
- Drawable represents anything that can be drawn to the screen, and is a
generic class to be extended by anything that will be drawn to the
screen ultimately
- Image is a generic class to be used for drawing a simple image to the
screen. Text is a generic class to be used for drawing simple text to
the screen.
- Clickable is an extension of Drawable that can perform an action when
clicked on, and is itself abstract to be extended by other classes for
anything clickable in the program
- Button is used as the most basic clickable to be a simple button (with
several size and color options) that has a clickable action associated
with it
- Draw is a util class to be used by Drawables to draw themselves to the
screen
- TextInput and its classes are in a very early state, but the point is
anything implementing TextInput allows the user to type something, and
that ultimately any text input by the user is done through a TextInput
object